Refactor marks pages into own file
